% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mangow-package.r, R/mangow.R
\docType{package}
\name{mangow}
\alias{mangow}
\alias{mangow-package}
\title{mangow.}
\usage{
mangow(data)
}
\arguments{
\item{data}{The data}
}
\description{
mangow.

Convert a Gower distance problem to a Manhattan distance problem
}

