% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangow.R
\name{mangow}
\alias{mangow}
\title{Convert a Gower distance problem to a Manhattan distance problem}
\usage{
mangow(data)
}
\arguments{
\item{data}{The input data frame
with continuous, categorical (factor) and ordered variables}
}
\value{
A numeric matrix
}
\description{
This function converts a data frame to a matrix with the same number of rows.
For any two rows of the resulting matrix, the Manhattan distance
equals the Gower distance in the input data between the corresponding rows.
}
\examples{
iris_sub <- iris[c(1:2,50:51,100:101), ]
row.names(iris_sub) <- NULL
iris_sub
cluster::daisy(iris_sub, "gower")
mangow_iris_sub <- mangow(iris_sub)
cluster::daisy(mangow_iris_sub, "manhattan")
}
